+++ include/distortos/OnceFlag.hpp
@@ -2,7 +2,7 @@
  * \file
  * \brief OnceFlag class header
  *
- * \author Copyright (C) 2015 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
+ * \author Copyright (C) 2015-2017 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
  *
  * \par License
  * This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of the MPL was not
@@ -12,9 +12,7 @@
 #ifndef INCLUDE_DISTORTOS_ONCEFLAG_HPP_
 #define INCLUDE_DISTORTOS_ONCEFLAG_HPP_
 
-#include "distortos/internal/synchronization/CallOnceControlBlock.hpp"
-
-#if DISTORTOS_CALLONCE_SUPPORTED == 1 || DOXYGEN == 1
+#include "distortos/Mutex.hpp"
 
 namespace distortos
 {
@@ -24,8 +22,6 @@ namespace distortos
  *
  * Similar to std::once_flag - http://en.cppreference.com/w/cpp/thread/once_flag
  * Similar to POSIX pthread_once_t - http://pubs.opengroup.org/onlinepubs/9699919799/functions/pthread_once.html#
- *
- * \note This class requires GCC 4.9.
  *
  * \ingroup synchronization
  */
@@ -42,19 +38,21 @@ public:
 	 */
 
 	constexpr OnceFlag() :
-			callOnceControlBlock_{}
+			mutex_{},
+			done_{}
 	{
 
 	}
 
 private:
 
-	/// internal internal::CallOnceControlBlock object used by callOnce()
-	internal::CallOnceControlBlock callOnceControlBlock_;
+	/// internal mutex used by callOnce()
+	Mutex mutex_;
+
+	/// tells whether any function was already called for this object (true) or not (false)
+	volatile bool done_;
 };
 
 }	// namespace distortos
-
-#endif	// DISTORTOS_CALLONCE_SUPPORTED == 1 || DOXYGEN == 1
 
 #endif	// INCLUDE_DISTORTOS_ONCEFLAG_HPP_