+++ include/distortos/ThreadState.hpp
@@ -2,7 +2,7 @@
  * \file
  * \brief ThreadState enum class header
  *
- * \author Copyright (C) 2015 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
+ * \author Copyright (C) 2015-2017 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
  *
  * \par License
  * This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of the MPL was not
@@ -43,8 +43,6 @@ enum class ThreadState : uint8_t
 	blockedOnConditionVariable,
 	/// thread is waiting for signal
 	waitingForSignal,
-	/// thread is blocked on OnceFlag
-	blockedOnOnceFlag,
 	/// internal thread object was detached
 	detached,
 };